Refactored io base
