+++ tsdat/io/base/input_key.py
@@ -1,0 +1,1 @@
+InputKey = str
