+++ tsdat/io/base/retrieved_variable.py
@@ -1,0 +1,18 @@
+from typing import (
+    List,
+    Union,
+)
+
+from pydantic import BaseModel, Extra
+
+from .data_converter import DataConverter
+from .input_key import InputKey
+
+
+# TODO: This needs a better name
+class RetrievedVariable(BaseModel, extra=Extra.forbid):
+    """Tracks the name of the input variable and the converters to apply."""
+
+    name: Union[str, List[str]]
+    data_converters: List[DataConverter] = []
+    source: InputKey = ""
