+++ tsdat/io/base/var_name.py
@@ -1,0 +1,1 @@
+VarName = str
